+++ classy-prelude/ChangeLog.md
@@ -1,6 +1,7 @@
 ## 1.0.0
 
 * Support for mono-traversable-1.0.0
+* Switch to safe-exceptions
 
 ## 0.12.8